create multimodule minimal Maven project
this project illustrate how to simplify use of the DAMapping annotation processor on a per-module basis by configuring the maven-compiler-plugin in the paren POM but declaring the DAMapping annotation processor dependency with the 'provided' scope in dependencyManagement. Module that will use DAMapping annotation processor will simply declare it as a direct dependency but will override the scope from dependencyManagement to 'compile'.